15 спринт 1 итерация

+++ src/main/java/ru/practicum/shareit/user/UserController.java
@@ -23,17 +23,17 @@ public class UserController {
     }
 
     @PatchMapping("/{userId}")
-    public UserDto updateUser(@PathVariable int userId, @RequestBody UserDto userDto) {
+    public UserDto updateUser(@PathVariable Long userId, @RequestBody UserDto userDto) {
         return userService.updateUser(userId, userDto);
     }
 
     @GetMapping("/{userId}")
-    public UserDto getUserById(@PathVariable int userId) {
+    public UserDto getUserById(@PathVariable Long userId) {
         return userService.getUserByID(userId);
     }
 
     @DeleteMapping("/{userId}")
-    public void deleteUserByID(@PathVariable int userId) {
+    public void deleteUserByID(@PathVariable Long userId) {
         userService.deleteUserByID(userId);
     }
